🔨 Refactoring: 내부에서 Error, Loading 처리 해주기
